Move old tests to basic_test.go
